Fixed deletion of lines and invoice line incrementing
Delete button is a little buggy - have to click it twice the first time
you try to delete a line after the page is rendered, but after that it
works fine, not sure why yet.

Also, removed the invoice line number from the page, and created a
variable that is incremented with each invoice line insertion, so that
it still give a line number, but I don’t have to dynamically calculate
it on the page.