"fixed" manufacturer dropdown issue, with auto publish
For some reason the Manufacturers collection isn’t being subscribed to
correctly, and I can’t figure out why. All the drop downs function as
expected when auto publish is installed added. When i removed it I
subscribed to all the data via individual pub/sub’s so not sure what
the issue was. Fine for now, but need to find the problem eventually.