Enabled CIU export of invoice lines.
Man this was a real b*&^%. Ultimately, I was solving the problem the
right way from the start, but the problem was that I kept using ‘i’ as
my loop counters, and when I passed the loop to export the invoice
lines to the prependZeros and appendSpaces functions, it started
kicking out of the loop and counting to i really high. Once i changed
the incrementer to ‘k’ everything was pretty much fine.

Also added meteorhacks:npm package to try and mitigate the ‘not-found’
error chrome keeps giving me. It seems to have fixed it so that I don’t
get the error when clicking the submit button, but it still will say
“not-defined” and point me to where fs is declared (i.e. fs =
Meteor.npmRequire(‘fs’)). Weird.