More robust invoice submission.
- Move invoiceLine insert code before the invoice update, like it used
to be: now invoice lines are inserted after initial invoice is
inserted, but before we update the invoice with additional properties.

- Added some or guards ( ‘ || 0 ‘ ) to the lineRetailCost variable, to
compensate for times when nothing is entered. the parseFloat plus
numeralJS formatting was causing an error when the lineRetailCost field
wasn’t filled in.

- Added another check in the InvoiceLine schema for the lineRetailCost
to make sure that it sets the value if it has no value when insert is
called.

- Lots of additional checks on Invoice schema. Probably overkill, but
it will at least make it easier to make future modifications to checks.